Fix assets
