+++ ckanext/og_datatablesview/assets/og_datatablesview.js
@@ -1,4 +1,5 @@
-var run_query = function(params, format) {
+// helper for filtered downloads
+var run_query = function (params, format) {
   var form = $('#filtered-datatables-download');
   var p = $('<input name="params" type="hidden"/>');
   p.attr("value", JSON.stringify(params));
@@ -7,9 +8,11 @@ var run_query = function(params, format) {
   f.attr("value", format);
   form.append(f);
   form.submit();
+  p.remove();
+  f.remove();
 }
 
-//main
+// main
 this.ckan.module('datatables_view', function (jQuery) {
   return {
     initialize: function() {
@@ -54,32 +57,3 @@ this.ckan.module('datatables_view', function (jQuery) {
     }
   }
 });
-
-// shake animation
-function animateEl(element, animation, complete) {
-    if (!element instanceof jQuery || !$(element).length || !animation) return null;
-
-    if (element.data('animating')) {
-        element.removeClass(element.data('animating')).data('animating', null);
-        element.data('animationTimeout') && clearTimeout(element.data('animationTimeout'));
-    }
-
-    element.addClass('animated-' + animation).data('animating', 'animated-' + animation);
-    element.data('animationTimeout', setTimeout((function() {
-        element.removeClass(element.data('animating')).data('animating', null);
-        complete && complete();
-    }), 400));
-}
-
-// custom error handler instead of default datatable alert error
-// this often happens when invalid datastore_search queries are returned
-$.fn.dataTable.ext.errMode = 'none';
-$('#dtprv').on('error.dt', function(e, settings, techNote, message) {
-    console.log('DataTables error: ', message);
-
-    // if error code 7, most probably an FTS query error. shake input
-    if (techNote == 7) {
-        const shake_element = $(":focus");
-        animateEl(shake_element, 'shake');
-    }
-})
