+++ ckanext/og_datatablesview/assets/resource.config
@@ -12,6 +12,8 @@ main =
     vendor/Select-1.2.2/css/select.bootstrap.css
     vendor/ColReorder-1.4.1/css/colReorder.bootstrap.css
 
+    og_datatablesview.css
+
     vendor/Bootstrap-3.3.7/js/bootstrap.js
     vendor/DataTables-1.10.15/js/jquery.dataTables.js
     vendor/DataTables-1.10.15/js/dataTables.bootstrap.js
@@ -28,4 +30,4 @@ main =
     vendor/Select-1.2.2/js/dataTables.select.js
     vendor/ColReorder-1.4.1/js/dataTables.colReorder.js
 
-    datatablesview.js
+    og_datatablesview.js
