+++ ckanext/og_datatablesview/assets/webassets.yml
@@ -1,5 +1,6 @@
 main-css:
-  output: ogdatatablepreview/%(version)s_main.css
+  output: ckanext-og_datatablesview/%(version)s_main.css
+  filters: cssrewrite
   contents:
     - vendor/Bootstrap-3.3.7/css/bootstrap.css
     - vendor/DataTables-1.10.15/css/dataTables.bootstrap.css
@@ -10,9 +11,11 @@ main-css:
     - vendor/Responsive-2.1.1/css/responsive.bootstrap.css
     - vendor/Select-1.2.2/css/select.bootstrap.css
     - vendor/ColReorder-1.4.1/css/colReorder.bootstrap.css
+    - og_datatablesview.css
 
 main-js:
-  output: ogdatatablepreview/%(version)s_main.js
+  output: ckanext-og_datatablesview/%(version)s_main.js
+  filters: rjsmin
   extra:
     preload:
       - base/main
@@ -33,4 +36,4 @@ main-js:
     - vendor/Responsive-2.1.1/js/dataTables.responsive.js
     - vendor/Select-1.2.2/js/dataTables.select.js
     - vendor/ColReorder-1.4.1/js/dataTables.colReorder.js
-    - datatablesview.js
+    - og_datatablesview.js
