+++ ckanext/og_datatablesview/plugin/flask_plugin.py
@@ -1,12 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from flask import Blueprint
-
 import ckan.plugins as p
-import ckanext.og_datatablesview.utils as utils
-
-
-ogdatatablesview = Blueprint(u'ogdatatablesview', __name__)
+import ckanext.og_datatablesview.blueprint as blueprint
 
 
 class MixinPlugin(p.SingletonPlugin):
@@ -14,21 +9,4 @@ class MixinPlugin(p.SingletonPlugin):
 
     # IBlueprint
     def get_blueprint(self):
-        return ogdatatablesview
-
-
-def ajax(resource_view_id):
-    return utils.ajax(resource_view_id)
-
-def filtered_download(resource_view_id):
-    return utils.filtered_download(resource_view_id)
-
-
-ogdatatablesview.add_url_rule(
-    u'/datatables/ajax/<resource_view_id>',
-    view_func=ajax, methods=[u'POST']
-)
-ogdatatablesview.add_url_rule(
-    u'/datatables/filtered-download/<resource_view_id>',
-    view_func=filtered_download, methods=[u'POST']
-)
+        return blueprint.ogdatatablesview
